Add full date to Notebook cards

+++ src/libs/utils.js
@@ -78,7 +78,7 @@ export const makePrettyDate = function(dateString) {
 
 export const makeCompleteDate = dateString => new Date(dateString)
   .toLocaleString(navigator.language,
-    { day: 'numeric', month: 'numeric', year: 'numeric', hour: 'numeric', minute: 'numeric' }
+    { day: 'numeric', month: 'short', year: 'numeric', hour: 'numeric', minute: 'numeric' }
   )
 
 export const formatUSD = new Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format
